+++ public/mkturk/index.js
@@ -1,8 +1,8 @@
 // Delete imports after developing. Solely for type checking pursposes.
-import firebase from 'firebase/app';
-import 'firebase/auth';
-import 'firebase/storage';
-import { formatWithOptions } from 'util';
+// import firebase from 'firebase/app';
+// import 'firebase/auth';
+// import 'firebase/storage';
+// import { formatWithOptions } from 'util';
 /* 
  * Check for MTurk tokens in the URL
  */
@@ -250,13 +250,33 @@ function skipHardwareDevice(event) {
   waitforClick.next(1);
 }
 
-(async () => {
-  document.querySelector('button[id=quickload]')
+
+(async function(){
+	document.querySelector('button[id=quickload]')
     .addEventListener('pointerup', quickLoad_listener, false);
 
   //--- for Safari
   document.querySelector('button[id=quickload')
     .addEventListener('click', quickLoad_listener, false);
+
+	// if ( ENV.WebUSBAvailable ){
+	// 	await usb_scriptLoaded
+	// 	document.querySelector ("button[id=connectusb]").addEventListener(
+	// 		'pointerup',findUSBDevice,false)
+	// 	document.querySelector("button[id=nousb]").addEventListener(
+	// 		'pointerup',skipHardwareDevice,false)
+	// 	document.querySelector("button[id=preemptRFID]").addEventListener(
+	// 		'pointerup',preemptRFID_listener,false)	
+
+	// 		//---- for Safari
+	// 		document.querySelector ("button[id=connectusb]").addEventListener(
+	// 			'click',findUSBDevice,false)
+	// 		document.querySelector("button[id=nousb]").addEventListener(
+	// 			'click',skipHardwareDevice,false)
+	// 		document.querySelector("button[id=preemptRFID]").addEventListener(
+	// 			'click',preemptRFID_listener,false)	
+	// 		//---- (END) for Safari
+  // }
 
   if (ENV.WebUSBAvailable) {
     await usb_scriptLoaded;
@@ -277,7 +297,7 @@ function skipHardwareDevice(event) {
 		//---- (END) for Safari
   }
 
-  if (ENV.WebBluetoothAvailable) {
+	if (ENV.WebBluetoothAvailable) {
 		await ble_scriptLoaded;
 		await blescale_scriptLoaded;
 		//Button callback for asynchronous connection to bluetooth scale
@@ -289,6 +309,26 @@ function skipHardwareDevice(event) {
       .addEventListener('click', blescaleconnect, false);
 		//---- (END) for Safari
   }
+
+	// document.querySelector("button[id=doneEditingParams]").addEventListener(
+	// 	'pointerup',doneEditingParams_listener,false)
+	// document.querySelector("button[id=doneTestingTask]").addEventListener(
+	// 	'pointerup',doneTestingTask_listener,false)
+	// document.querySelector("button[id=stressTest]").addEventListener(
+	// 	'touchstart',stressTest_listener,false)
+	// document.querySelector("button[id=gridPoints]").addEventListener(
+	// 	'touchstart',gridPoints_listener,false)
+
+	// 	//---- for Safari
+	// 	document.querySelector("button[id=doneEditingParams]").addEventListener(
+	// 		'click',doneEditingParams_listener,false)
+	// 	document.querySelector("button[id=doneTestingTask]").addEventListener(
+	// 		'click',doneTestingTask_listener,false)
+	// 	document.querySelector("button[id=stressTest]").addEventListener(
+	// 		'click',stressTest_listener,false)
+	// 	document.querySelector("button[id=gridPoints]").addEventListener(
+	// 		'click',gridPoints_listener,false)
+  // 	//---- (END) for Safari
   
   document.querySelector("button[id=doneEditingParams]")
     .addEventListener('pointerup', doneEditingParams_listener, false);
@@ -309,22 +349,23 @@ function skipHardwareDevice(event) {
   document.querySelector("button[id=gridPoints]")
     .addEventListener('click', gridPoints_listener, false);
   //---- (END) for Safari
-  
-  //====================== Retrieve device's screen properties ===========================//
-	ENV.UserAgent = window.navigator.userAgent;
-	ENV.DeviceScreenWidth = window.screen.width;
-	ENV.DeviceScreenHeight = window.screen.height;
 
-	var deviceProperties = await deviceDetect();
-	ENV.DeviceType = deviceProperties.data.device.type;
-	ENV.DeviceBrand = deviceProperties.data.device.brand;
-	ENV.DeviceName = deviceProperties.data.device.model;
-	ENV.DeviceGPU = deviceProperties.data.gpu.renderer;
-	ENV.DeviceBrowserName = deviceProperties.data.client.name;
-	ENV.DeviceBrowserVersion = deviceProperties.data.client.version;
-	ENV.DeviceOSName = deviceProperties.data.os.name;
-	ENV.DeviceOSVersion = deviceProperties.data.os.version;
-	ENV.DeviceTouchscreen = deviceProperties.data.touchscreen;
+
+	//====================== Retrieve device's screen properties ===========================//
+	ENV.UserAgent = window.navigator.userAgent
+	ENV.DeviceScreenWidth = window.screen.width
+	ENV.DeviceScreenHeight = window.screen.height
+
+	var deviceProperties = await deviceDetect()
+	ENV.DeviceType = deviceProperties.data.device.type
+	ENV.DeviceBrand = deviceProperties.data.device.brand
+	ENV.DeviceName = deviceProperties.data.device.model
+	ENV.DeviceGPU = deviceProperties.data.gpu.renderer
+	ENV.DeviceBrowserName = deviceProperties.data.client.name
+	ENV.DeviceBrowserVersion = deviceProperties.data.client.version
+	ENV.DeviceOSName = deviceProperties.data.os.name
+	ENV.DeviceOSVersion = deviceProperties.data.os.version
+	ENV.DeviceTouchscreen = deviceProperties.data.touchscreen
 
 	var screenSpecs = await queryDeviceonFirestore(ENV.DeviceName);
 	//if device not identified by deviceAPI or no matching firestore devices record found for an identified device
@@ -373,12 +414,12 @@ function skipHardwareDevice(event) {
 		event.type = "AutoConnect";
 		await findUSBDevice(event);
   }
-  
-  //====================== Quickload Button Set-up ===========================//
+
+	//====================== Quickload Button Set-up ===========================//
 	// GET PARAMFILE NAME
   var subjectlistobj = document.getElementById("subjectID_select");
 
-  for (let i = subjectlist.length; i > 0; i--) {
+  for (let i = subjectlist.length - 1; i >= 0; i--) {
     let opt = document.createElement('option');
     opt.value = i;
     opt.innerHTML = subjectlist[i];
@@ -389,7 +430,8 @@ function skipHardwareDevice(event) {
   subjectlistobj.style.visibility = "visible";
   
   if (localStorage.getItem('Agent') != null) { // IF agent stored locally, show quickload button
-    QuickLoad.agent = localStorage.getItem('ConnectUSB');
+    QuickLoad.agent = localStorage.getItem('Agent');
+    QuickLoad.connectusb = localStorage.getItem('ConnectUSB');
 
     if (QuickLoad.connectusb == null) {
       QuickLoad.connectusb = 0;
@@ -404,10 +446,10 @@ function skipHardwareDevice(event) {
 			document.querySelector("button[id=quickload]").innerHTML = QuickLoad.agent + " <i>USB</i>";
 		}
   } else { // ELSE don't show button
-    document.querySelector('button[id=quickload]').style.display = 'none';
+    document.querySelector("button[id=quickload]").style.display = 'none';
   }
-  //====================== (END) Quickload Set-up ===========================//
-  
+	//====================== (END) Quickload Set-up ===========================//
+
   //================== AWAIT LOAD SUBJECT PARAMS ==================//
 	document.querySelector("div[id=subjectID_div]").style.display = "block";
 	document.querySelector("div[id=subjectID_div]").style.visibility = "visible";
@@ -424,10 +466,43 @@ function skipHardwareDevice(event) {
 	let rtdbAgentConnectionRef = rtdb.ref(`agents/${ENV.Subject}/numConnections`);
 	FLAGS.rtdbDataRef = rtdb.ref('data/' + ENV.Subject);
   //================== (END) AWAIT LOAD SUBJECT PARAMS ==================//
-  
-  //====================== Connect USB ===========================//
-	if (ENV.WebUSBAvailable) {
 
+
+	//====================== Connect USB ===========================//
+	// if ( ENV.WebUSBAvailable ){
+
+	// 	if (typeof(port.connected) == 'undefined' || port.connected == false){
+	// 		var event = {}
+	// 		event.type = "AutoConnect"
+	// 		await findUSBDevice(event)
+	// 	}
+
+	// 	if ( (typeof(port.connected) == 'undefined' || port.connected == false) &&
+	// 	     (QuickLoad.load == 0 || (QuickLoad.load == 1 && QuickLoad.connectusb == 1))){
+	// 		//=============== AWAIT CONNECT TO HARDWARE (via USB) ===============//
+	// 		port.connected = false
+	// 		document.querySelector("button[id=connectusb]").style.display = "block"
+	// 		document.querySelector("button[id=connectusb]").style.visibility = "visible"
+	// 		document.querySelector("button[id=nousb]").style.display = "block"
+	// 		document.querySelector("button[id=nousb]").style.visibility = "visible"
+
+	// 		await connectHardwareButtonPromise()
+	// 	} //IF !QuickLoad.load
+
+	// 	document.querySelector("button[id=connectusb]").style.display = "none"
+	// 	document.querySelector("button[id=nousb]").style.display = "none"
+	// }
+	// else {
+	// 	//skip usb device connection
+	// 	port={
+	// 	  statustext_connect: "",
+	// 	  statustext_sent: "",
+	// 	  statustext_received: "",
+	// 	  connected: false
+	// 	}
+  // }
+  
+	if (ENV.WebUSBAvailable) {
 		if (
       typeof(port.connected) == 'undefined'
       || port.connected == false
@@ -465,1229 +540,6 @@ function skipHardwareDevice(event) {
 		  statustext_received: "",
 		  connected: false
 		};
-	}
-  //====================== (END) Connect USB ===========================//
-  
-  if (ENV.WebBluetoothAvailable == 0) {
-		blescale = {
-			connected: 0,
-			statustext_connect: "",
-			statustext_sent: "",
-			statustext_received: "",
-    };
-    
-		ble = {
-			connected: 0,
-		};
-	}
-
-	//================== AWAIT USER CAN EDIT SUBJECT PARAMS ==================//
-	if (QuickLoad.load == 0) {
-		updateStatusText(JSON.stringify(TASK, null, ' '));
-    document.querySelector("p[id=headsuptext]").setAttribute("contentEditable", true);
-		document.querySelector("button[id=doneEditingParams]").style.display = "block";
-		document.querySelector("button[id=doneEditingParams]").style.visibility = "visible";
-
-		await editParamsPromise();
-		document.querySelector("button[id=doneEditingParams]").style.display = "none";
-		var textobj = document.getElementById("headsuptext");
-		textobj.removeEventListener('touchend', headsuptext_listener);
-		textobj.removeEventListener('mouseup', headsuptext_listener);
-    document.querySelector("p[id=headsuptext]").setAttribute("contentEditable", false);
-
-		if (FLAGS.need2saveParameters == 1) {
-			var user_param_text = document.getElementById("headsuptext").innerHTML; //get new params
-			await saveParameterTexttoFirebase(user_param_text); //write new params
-			await loadParametersfromFirebase(ENV.ParamFileName); //then read them
-		} //IF 
-	} //IF !QuickLoad.load
-  //================== (END) AWAIT USER CAN EDIT SUBJECT PARAMS ==================//
-  
-  // =================== LOAD MKMODELS IF SPECIES = MODEL =================//
-	let mkm;
-	if (TASK.Species == 'model') {
-		mkm = new MkModels();
-		let fromTFHub = TASK.ModelConfig.modelURL.includes('tfhub');
-		await mkm.loadFeatureExtractor(TASK.ModelConfig.modelURL, { fromTFHub: fromTFHub });
-		let cvs = document.getElementById('model-canvas');
-		mkm.bindCanvasElement(cvs);
-		mkm.buildClassifier(TASK.ModelConfig);
-	}
-  // ======================== (END) LOAD MKMODELS ========================//
-  
-
-  //============= AWAIT READ SUBJECT PERFORMANCE HISTORY =============//
-	// Read performance history
-	var subject_behavior_save_directory = DATA_SAVEPATH + ENV.Subject + '/';
-	if (TASK.Automator != 0) {
-		var history_file_paths = (
-      await getMostRecentBehavioralFilePathsFromFirebase(
-        ndatafiles2read,
-        ENV.Subject,
-        subject_behavior_save_directory,
-      )
-    );
-		trialhistory = await readTrialHistoryFromFirebase(history_file_paths);		
-	}
-
-	//===================== AWAIT INITIALIZE AUTOMATOR =================//
-	// Initialize automator - change TASK to that specified by TASK.CurrentAutomatorStage. 
-	var num_prebuffer_trials = 200;
-	if (TASK.Automator != 0) {
-		automator_data = await loadTextfromFirebase(TASK.AutomatorFilePath);
-		automateTask(automator_data, trialhistory);
-		await saveParameterstoFirebase();
-		await loadParametersfromFirebase(ENV.ParamFileName);
-	}//IF TASK.Automator != 0
-
-	//============= AWAIT LOAD SOUNDS =============//
-	soundpromises = sounds.serial.map(loadSoundfromFirebase); //create array of sound load Promises
-	await Promise.all(soundpromises); //simultaneously evaluate array of sound load promises
-	updateStatusText("");
-
-  //============= AWAIT ESTIMATE SCREEN REFRESH RATE =========//
-  var fps = await estimatefps();
-  ENV.FrameRateDisplay = fps;
-  ENV.FrameRateMovie = fps/2;
-
-	//========= Start in TEST mode =======//
-	document.querySelector("button[id=googlesignin]").style.display = "none"; //if do style.visibility=hidden, element will still occupy space
-	document.querySelector("button[id=reloadpage]").style.display = "block";
-	document.querySelector("button[id=reloadpage]").style.visibility = "visible";
-
-	document.querySelector("button[id=doneTestingTask]").style.display = "block";
-	document.querySelector("button[id=doneTestingTask]").style.visibility = "visible";
-	document.querySelector("button[id=gridPoints]").style.display = "block";
-	document.querySelector("button[id=gridPoints]").style.visibility = "visible";
-
-	FLAGS.need2loadParameters = 1;
-	FLAGS.need2loadScenes = 1;
-	CURRTRIAL.num = 0;
-	EVENTS.trialnum = 0;
-  FLAGS.savedata = 0; // test trials can be performed, but data won't be saved
-  
-
-  // =========================================================================================================== // 
-  // ============ MAIN LOOP ==================================================================================== // 
-  // =========================================================================================================== // 
-  while (true) {
-    if (FLAGS.need2loadParameters == 1) {
-      FLAGS.need2loadParameters = await loadParametersfromFirebase(ENV.ParamFileName);
-  
-      if (TASK.Agent == "SaveImages") { 
-        document.querySelector("button[id=stressTest]").innerHTML = "Save Images";
-        TASK.SamplingStrategy = "sequential";
-        console.log("Automatically using sequential sampling since SAVE IMAGES was specified.");
-      }//IF SaveImages
-  
-      // TODO: if (typeof(TASK.DragtoRespond == undefined))
-      // if (typeof(TASK.DragtoRespond) != "undefined") { // IF defined, do nothing
-      // } else {
-      //   if (FLAGS.trackeye == 0) { //IF touch, then only clicking
-      //     TASK.DragtoRespond = 0; //click in box
-      //   } else if (FLAGS.trackeye != 0){ //ELSE IF eyetracker, allow dragging
-      //     TASK.DragtoRespond = 1 //drag into box
-      //   }
-      // }//ELSE default to no eye tracking
-
-      // TODO: Changed the code above to the following
-      if (typeof(TASK.DragtoRespond) == 'undefined') { // IF undefined, default to no eye tracking
-        if (FLAGS.trackeye) { // IF eyetracker, allow dragging
-          TASK.DragtoRespond = 1; // drag into box
-        } else { // IF touch (not eyetracking), then only clicking
-          TASK.DragtoRespond = 0;
-        }
-      }
-  
-      //load previous calibration if available
-      if (FLAGS.trackeye > 0){ // IF trackeye
-        //Calibration
-        ENV.Eye.calibration = 0;
-        ENV.Eye.CalibXTransform = [];
-        ENV.Eye.CalibYTransform = [];
-        ENV.Eye.CalibType = 'default';
-        ENV.Eye.NCalibPointsTrain = 0;
-        ENV.Eye.NCalibPointsTest = 0;
-        ENV.Eye.CalibTrainMSE = [];
-        ENV.Eye.CalibTestMSE = [];
-  
-        await loadEyeCalibrationfromFirestore(ENV.Subject);
-
-        if (ENV.Eye.CalibXTransform.length == 0) { // Default calibration
-  
-          var xrange = 0.5;
-          var yrange = 0.5;
-          var xscale = ENV.ViewportPixels[0] / xrange;
-          var yscale = ENV.ViewportPixels[1] / yrange;
-  
-          ENV.Eye.CalibXTransform = [xscale, 0, -(0.5 - xrange/2)*xscale]
-          ENV.Eye.CalibYTransform = [
-            0,
-            -yscale,
-            ENV.ViewportPixels[1] + (0.5 - yrange / 2 ) * yscale,
-          ];
-  
-          // ENV.Eye.CalibXTransform = [ 1, 0, 0]
-          // ENV.Eye.CalibYTransform = [ 0, 1, 0]
-  
-          saveEyeCalibrationtoFirestore(
-            ENV.Eye.CalibXTransform,
-            ENV.Eye.CalibYTransform,
-            ENV.Eye.NCalibPoints,
-            ENV.Eye.CalibType,
-          );
-
-        }
-
-        // will calibrate using TASK.CalibrateEye number of trials for train & same number for test
-        if (TASK.CalibrateEye > 0) { // IF CalibrateEye
-          ENV.Eye.calibration = 1;
-          ENV.Eye.NCalibPointsTrain = 0;
-          ENV.Eye.NCalibPointsTest = 0;
-          ENV.Eye.CalibTrainMSE = [];
-          ENV.Eye.CalibTestMSE = [];
-        }
-
-      }
-  
-      //============= SET UP CANVAS =============//
-      // Update canvas based on latest TASK state: 
-      refreshCanvasSettings(TASK); 
-      setupCanvasHeadsUp();
-      setupImageLoadingText();
-      windowWidth = document.body.clientWidth; //get true window dimensions at last possible moment
-      windowHeight = document.body.clientHeight;
-      setupCanvas(VISIBLECANVAS);
-  
-       //Foreground canvas that displays eye position during practice screen
-      setupEyeTrackerCanvas();
-  
-      if (ENV.DevicePixelRatio !== 1) {
-        scaleCanvasforHiDPI(VISIBLECANVAS);
-        scaleCanvasforHiDPI(EYETRACKERCANVAS);
-      }
-      
-      if (ENV.OffscreenCanvasAvailable) {
-        OFFSCREENCANVAS = new OffscreenCanvas(VISIBLECANVAS.width, VISIBLECANVAS.height);
-        OFFSCREENCANVAS.commitTo = (dest) => {
-          try {
-            let bitmap = this.transferToImageBitmap();
-            dest.transferFromImageBitmap(bitmap);
-            return { status: 'succeeded' };
-          } catch (e) {
-            console.error('[OFFSCREENCANVAS.commitTo] Error:', e);
-            return { status: 'failed' };
-          }
-        }
-        // OFFSCREENCANVAS.commitTo = function(dest) {
-        //   try {
-        //     var bitmap = this.transferToImageBitmap()
-        //     dest.transferFromImageBitmap(bitmap)
-        //     str = {status: "succeeded"}
-        //     return str
-        //   }
-        //   catch(error){
-        //     console.log(error)
-        //     str = {status: "failed"}
-        //     return str				
-        //   }
-        // }
-      } else {
-        OFFSCREENCANVAS = VISIBLECANVAS;
-      }
-  
-      CANVAS.workspace = [
-        0,
-        0,
-        VISIBLECANVAS.width,
-        VISIBLECANVAS.height,
-      ];
-  
-      TQS = undefined;
-      FLAGS.need2loadScenes = 1; 
-  
-      //Determine task type
-      if (TASK.RewardStage == 0) {
-        ENV.Task = "FIXATION"
-      } else if (TASK.RewardStage == 1) { // IF Task.RewardStage
-        if (TASK.NRSVP > 0) {
-          ENV.Task = 'RSVP';
-        } else if (TASK.SameDifferent > 0 && TASK.ChoiceGridIndex.length == 2) { // Task is SameDifferent
-          // Same-Different (SD)
-          ENV.Task = 'SD';
-        } else if (TASK.ObjectGridIndex.length == TASK.ImageBagsSample.length) { // Task is Stimulus-Response
-          // Stimulus-Response (SR)
-          ENV.Task = 'SR';
-        } else { // Task is Match-to-Sample
-          // Match-to-Sample
-          ENV.Task = 'MTS';
-        }
-      }
-  
-      //Size of Fixation screen circle or image
-      ENV.FixationRadius = TASK.FixationSizeInches / 2 * ENV.ViewportPPI;
-  
-      //Size of Choice screen circle or square
-      ENV.ChoiceRadius = TASK.ChoiceSizeInches / 2 * ENV.ViewportPPI;
-  
-      //Fixation dot, if >0, will appear on both fixation & sample screens
-      ENV.FixationDotRadius = TASK.FixationDotSizeInches / 2 * ENV.ViewportPPI;
-  
-      //Fixation window, if specified, operates on both fixation & sample screens
-      ENV.FixationWindowRadius = TASK.FixationWindowSizeInches / 2 * ENV.ViewportPPI;
-  
-      // define image display grid
-      funcreturn = defineImageGrid(
-        TASK.NGridPoints,
-        TASK.GridSpacingInches * ENV.ViewportPPI,
-        TASK.GridXOffsetInches * ENV.ViewportPPI,
-        TASK.GridYOffsetInches * ENV.ViewportPPI,
-      );
-      xcanvascenter = funcreturn[0];
-      ycanvascenter = funcreturn[1];
-      ENV.XGridCenter = funcreturn[2];
-      ENV.YGridCenter = funcreturn[3];
-  
-      FLAGS.purge = 1;
-      FLAGS.createnewfirestore = 1;
-      CURRTRIAL.reset();
-      EVENTS.reset_trialseries();
-      EVENTS.reset_timeseries();
-    } //IF need2loadParameters
-
-    if (FLAGS.purge == 1) {
-      purgeTrackingVariables();
-      FLAGS.purge = 0; 
-    } //IF purge
-
-
-    //======================== 3D SCENE SET-UP =======================//
-    if (FLAGS.need2loadScenes) {
-      IMAGES = { Sample: [], Test: [] };
-      IMAGEMETA = {};
-
-      // STEPS FOR 3D SCENE SET-UP
-	  	// ---- 0: load scene params from JSON
-		  // 0: expand trial params & get mesh paths
-		  // ---- 1: load meshes
-		  // ---- 2: init scene & camera
-		  // ---- 3: add all lights & objects
-		  // ---- 4: compile shaders
-		  // 5: select frame to render
-      // 5: animate <--> render loop within trial
-      
-
-      //============ 0: LOAD SCENES from JSON ============//
-      for (let i = 0; i < TASK.ImageBagsSample.length; i++) {
-        IMAGES.Sample[i] = await loadTextfromFirebase(TASK.ImageBagsSample[i]);
-      }
-
-      for (let i = 0; i < TASK.ImageBagsTest.length; i++) {
-        IMAGES.Test[i] = await loadTextfromFirebase(TASK.ImageBagsTest[i]);
-      }
-
-      // find the longest scene param arry in IMAGES (ie # of trials)
-      for (let i = 0; i < IMAGES.Sample.length; i++) {
-        IMAGES.Sample[i].nimages = getLongestArray(IMAGES.Sample[i]);
-        IMAGES.Test[i].nimages = getLongestArray(IMAGES.Test[i]);
-
-        //Determine if images will also be rendered
-			  IMAGES.Sample[i].nbackgroundimages = IMAGES.Sample[i].IMAGES.imageidx.length;
-        IMAGES.Test[i].nbackgroundimages = IMAGES.Test[i].IMAGES.imageidx.length;
-        
-        FLAGS.movieper['Sample'][i] = [];
-        FLAGS.movieper['Test'][i] = [];
-      }
-      //============ (END) 0: LOAD SCENES from JSON ============//
-
-      //============ 1: LOAD MESHES FOR SCENES ============//
-      OBJECTS = { Sample: {}, Test: {} };
-      for (let taskscreen in OBJECTS) {
-        let meshPaths = [];
-        let meshIdxs = [];
-
-        for (let classLabel = 0; classLabel < IMAGES[taskscreen].length; classLabel++) {
-          for (const obj in IMAGES[taskscreen][classLabel].OBJECTS) {
-            meshPaths.push(IMAGES[taskscreen][classLabel].OBJECTS[obj].meshpath);
-            meshIdxs.push([classLabel, obj]); 
-          }
-        }
-
-        let meshes = await loadMeshArrayfromFirebase(meshPaths);
-
-        // FOR i meshes, initialize corresponding label to an empty array
-        for (let i = 0; i < meshes.length; i++) {
-          let meshLabel = meshIdxs[i][0];
-          OBJECTS[taskscreen][meshLabel] = { meshes: [] };
-        }
-
-        // For i meshes, store in corresponding labels
-        for (let i = 0; i < meshes.length; i++) {
-          let meshLabel = meshIdxs[i][0];
-          let meshName = meshIdxs[i][1];
-          OBJECTS[taskscreen][meshLabel].meshes[meshName] = meshes[i];
-        }
-      }
-      //============ (END) 1: LOAD MESHES FOR SCENES ============//
-
-      //============ 2: INIT SCENE & CAMERA ============//
-      setupCanvas(VISIBLECANVASWEBGL);
-		  await initThreeJS(IMAGES);
-      //============ (END) 2: INIT SCENE & CAMERA ============//
-      
-
-      //============ 3: ADD ALL LIGHTS/OBJECTS TO SCENE ============//
-	    CAMERAS = { Sample: {}, Test: {} };
-	    LIGHTS = { Sample: {}, Test: {} };
-		  for (let scenetype in scene) {
-			  await addToScene(scenetype);
-		  } 
-		  console.log('3js: added lights & objects');
-      //============ (END) 3: ADD ALL LIGHTS/OBJECTS TO SCENE ============//
-      
-   		//============ 4: PRELOAD SHADERS (COMPILE) ============//
-      for (let scenetype in scene) {
-        renderer.compile(
-          scene[scenetype],
-          scene[scenetype].getObjectByName('cam0')
-        );
-      }
-      console.log('3js: compiled scene')
-   		//============ (END) 4: PRELOAD SHADERS (COMPILE) ============//
-
-      FLAGS.need2loadScenes = 0;
-
-      // Make a scene trial queue TQS (overrides TQ)
-      TQS = new TrialQueueScene(TASK.SamplingStrategy);
-      await TQS.build(num_prebuffer_trials);
-
-      // Store scene metadata
-      let sampleSceneMeta = (
-        objectomeSceneNamesToLatentVars(
-          TASK.ImageBagsSample,
-          TQS.testbag_labels,
-          IMAGES.Sample
-        )
-      );
-      let sampleSceneMetaKeys = Object.keys(sampleSceneMeta);
-      for (let i = 0; i < sampleSceneMetaKeys.length; i++) {
-        IMAGEMETA['Sample' + sampleSceneMetaKeys[i]] = sampleSceneMeta[sampleSceneMetaKeys[i]];
-      }
-
-      let testSceneMeta = (
-        objectomeSceneNamesToLatentVars(
-          TASK.ImageBagsTest,
-          TQS.testbag_labels,
-          IMAGES.Test
-        )
-      );
-      let testSceneMetaKeys = Object.keys(testSceneMeta);
-      for (let i = 0; i < testSceneMetaKeys.length; i++) {
-        IMAGEMETA['Test' + testSceneMetaKeys[i]] = testSceneMeta[testSceneMetaKeys[i]];
-      }
-    }
-
-    if (typeof(TASK.BackgroundColor2D) == 'undefined') {
-      TASK.BackgroundColor2D = '#7F7F7F';
-    }
-    document.body.style.background = TASK.BackgroundColor2D;
-    //========================(END) 3D SCENE SET-UP =======================//
-
-    //============ SELECT SAMPLE & TEST IMAGES ============//
-
-    let imgSeqLen = (
-      (typeof(TASK.NRSVP) == 'undefined' || TASK.NRSVP <= 0) ? 1 : TASK.NRSVP
-    );
-
-    for (let i = 0; i < imgSeqLen; i++) {
-      let x = await TQS.get_next_trial();
-      CURRTRIAL.sampleimage[i] = x[0];
-      CURRTRIAL.sampleindex[i] = x[1];
-
-      // Sample can have multiple sequential scenes (items are over time; eg, RSVP)
-      CURRTRIAL.sampleindex_nonarray[i] = x[1][0];
-      CURRTRIAL.sample_scenebag_label[i] = x[5];
-      CURRTRIAL.sample_scenebag_index[i] = x[6];
-
-      // Test can have multiple simultaneous scenes (items are over space; ev, MtS)
-      if (i == 0) { // IF first image
-        CURRTRIAL.testimages[i] = x[2];
-        CURRTRIAL.testindices[i] = x[3]
-        CURRTRIAL.test_scenebag_labels[i] = x[7]
-        CURRTRIAL.test_scenebag_indices[i] = x[8]
-        CURRTRIAL.correctitem = x[4]
-        samplereward = x[9]
-      }
-    }
-
-    logEVENTS("Sample", CURRTRIAL.sampleindex_nonarray, "trialseries");
-    logEVENTS("Test", CURRTRIAL.testindices[0], "trialseries");
-    //============(END) SELECT SAMPLE & TEST IMAGES ============//
-
-    //============ SET UP SAMPLE & TEST SEQUENCE ============//
-    // when & where to display
-    CURRTRIAL.tsequence = [0];
-    CURRTRIAL.sequencegridindex = [[-1]];
-
-    // what to display
-    CURRTRIAL.sequenceclip = [-1]; //movieclip# in RSVP
-    CURRTRIAL.sequenceframe = [-1]; //frame# in movie
-    CURRTRIAL.sequencetaskscreen = ['blank'];
-    CURRTRIAL.sequencelabel = [[0]]; //image class
-    CURRTRIAL.sequenceindex = [[0]]; //image index
-
-    //EXPAND SAMPLE (for rsvp & movies)
-    //Start with blank for max(100,SampleOFF), then append SampleON+blank (eg, blank,Sample,blank,Sample,blank)
-    for (let i = 0; i < CURRTRIAL.sample_scenebag_index.length; i++) { // FOR i RSVP Sample
-      let t0 = CURRTRIAL.tsequence[CURRTRIAL.tsequence.length - 1];
-      let sampleon = (
-        chooseArrayElement(
-          IMAGES['Sample'][CURRTRIAL.sample_scenebag_label[i][0]].durationMS,
-          CURRTRIAL.sample_scenebag_index[i][0],
-          0
-        )
-      );
-
-      // Timing: blankdurationpre, sampleon, framerate
-      let blankdurationpre = (
-        (i == 0) ? Math.max(100, TASK.SampleOFF) : TASK.SampleOFF
-      );
-
-      // Create Movie Sequence
-      [movie_sequence, movie_tsequence, movie_framenum] = (
-        createMovieSeq('Sample', blankdurationpre, sampleon, TASK.SampleOFF, ENV.FrameRateMovie)
-      );
-      movie_tsequence = (
-        movie_tsequence.map((a) => {
-          return a + t0;
-        })
-      );
-
-      CURRTRIAL.tsequence.push(...movie_tsequence);
-      CURRTRIAL.sequencegridindex.push(
-        ...Array(movie_tsequence.length).fill([TASK.SampleGridIndex])
-      );
-
-      CURRTRIAL.sequenceclip.push(
-        ...Array(movie_tsequence.length).fill(i)
-      );
-
-      CURRTRIAL.sequenceframe.push(...movie_framenum);
-      CURRTRIAL.sequencetaskscreen.push(...movie_sequence);
-      CURRTRIAL.sequencelabel.push(
-        ...Array(movie_tsequence.length).fill(CURRTRIAL.sample_scenebag_label[i])
-      );
-
-      CURRTRIAL.sequenceindex.push(
-        ...Array(movie_tsequence.length).fill(CURRTRIAL.sample_scenebag_index[i])
-      );
-    }
-
-    // APPEND TEST OR CHOICE
-    if (TASK.NRSVP <= 0) { // IF !RSVP, then show test/choice screen
-      let t0 = CURRTRIAL.tsequence[CURRTRIAL.tsequence.length - 1];
-      let teston = (
-        chooseArrayElement(
-          IMAGES["Test"][CURRTRIAL.test_scenebag_labels[0][0]].durationMS,
-          CURRTRIAL.test_scenebag_indices[0][0],
-          0,
-        )
-      );
-
-      if (typeof(teston) == 'undefined') {
-        console.log('Without this if, then print-to-console code, teston is undefined. Not clear why this strange behavior happens. Something to do with chooseArrayElement returning in time.');
-      }
-
-      [movie_sequence, movie_tsequence, movie_framenum] = (
-        createMovieSeq('Test', TASK.SampleOFF, teston, TASK.TestOFF, ENV.FrameRateMovie)
-      );
-      movie_tsequence = (
-        movie_tsequence.map((a) => {
-          return a + t0;
-        })
-      );
-
-      CURRTRIAL.tsequence.push(...movie_tsequence);
-      CURRTRIAL.sequencegridindex.push(
-        ...Array(movie_tsequence.length).fill(TASK.TestGridIndex)
-      );
-
-      CURRTRIAL.sequenceclip.push(
-        ...Array(movie_tsequence.length).fill(0)
-      );
-
-      CURRTRIAL.sequenceframe.push(...movie_framenum);
-      CURRTRIAL.sequencetaskscreen.push(...movie_sequence);
-      CURRTRIAL.sequencelabel.push(
-        ...Array(movie_tsequence.length).fill(CURRTRIAL.test_scenebag_labels[0])
-      );
-
-      CURRTRIAL.sequenceindex.push(
-        ...Array(movie_tsequence.length).fill(CURRTRIAL.test_scenebag_indices[0])
-      );
-
-      // Append choice if needed
-      if (TASK.SameDifferent > 0) { // IF Same-Different, show test & choice
-        let t0 = CURRTRIAL.tsequence[CURRTRIAL.tsequence.length - 1];
-        let seq;
-        let tseq;
-        
-        if (TASK.TestOFF > 0) {
-          seq = ['blank', 'choice'];
-          tseq = [t0, t0 + TASK.TestOFF];
-        } else {
-          seq = ['choice'];
-          tseq = [t0];
-        }
-
-        CURRTRIAL.tsequence.push(...tseq);
-        CURRTRIAL.sequencegridindex.push(
-          ...Array(tseq.length).fill(TASK.ChoiceGridIndex)
-        );
-
-        CURRTRIAL.sequenceclip.push(...Array(tseq.length).fill(0));
-        CURRTRIAL.sequenceframe.push(...Array(tseq.length).fill(0));
-        CURRTRIAL.sequencetaskscreen.push(...seq);
-        CURRTRIAL.sequencelabel.push(...Array(tseq.length).fill([0]));
-        CURRTRIAL.sequenceindex.push(...Array(tseq.length).fill([0]));
-      }
-
-    }
-    //============(END) SET UP SAMPLE & TEST SEQUENCE ============//
-
-    //================= RFID check =================//
-    /**
-     * If no matching read in the last TASK.CheckRFID seconds, wait for matching read
-     * (kicks-them-off model where they can work as long as reading, but then get 
-     * kicked off within TASK.CheckRFID seconds if they are the wrong agent or no reads)
-     */
-    if (
-      TASK.CheckRFID > 0
-      && ENV.AgentRFID != 'XX'
-      && FLAGS.savedata == 1
-    ) {
-      if (!port.connected) {
-        console.log('NO USB DEVICE CONNECTED: cannot check RFID!!');
-      } else {
-        let nreads = Object.keys(EVENTS['timeseries']['RFIDTag']).length;
-        // IF RFID does not check out, wait for a recent RFID read before proceeding with the next trial
-        if (
-          !(nreads > 0
-          && EVENTS['timeseries']['RFIDTag'][nreads - 1][2] == ENV.AgentRFID
-          && (Date.now() - new Date(EVENTS['timeseries']['RFIDTag'][nreads - 1][1])) < TASK.CheckRFID)
-        ) {
-          await rfid_promise(ENV.AgentRFID, TASK.CheckRFID);
-        } 
-      }
-    }
-    //================= (end) RFID check =================//
-
-
-    // FIXATION   FIXATION   FIXATION   FIXATION   FIXATION   FIXATION   //
-    // FIXATION   FIXATION   FIXATION   FIXATION   FIXATION   FIXATION   //
-    // FIXATION   FIXATION   FIXATION   FIXATION   FIXATION   FIXATION   //
-    //============ WHILE RUN FIXATION SCREEN ============//
-    FLAGS.waitingforTouches = TASK.NFixations;
-    if (TASK.RewardStage == 0) {
-      FLAGS.punishOutsideTouch = 1;
-    }
-
-    CURRTRIAL.allfixationxyt = [];
-    while (FLAGS.waitingforTouches > 0) {
-      // Choose fixation grid index at random
-      if (TASK.FixationGridIndex > 0) {
-        CURRTRIAL.fixationgridindex = TASK.FixationGridIndex;
-      } else if (TASK.FixationGridIndex < 0) {
-        CURRTRIAL.fixationgridindex = Math.floor(
-          Math.random() * ENV.XGridCenter.length
-        );
-      }
-      logEVENTS('FixationGridIndex', CURRTRIAL.fixationgridindex, 'trialseries');
-
-      if (TASK.FixationUsesSample <= 0) { // IF !FixationUsesSample, show fixation dot
-        // Render fixation screen
-        if (TASK.Species == 'macaque' || TASK.Species == 'human') {
-          ENV.FixationColor = 'white';
-        } else if (TASK.Species == 'marmoset') {
-          ENV.FixationColor = 'blue';
-        }
-        frame.shown = [];
-        frame.frames = [];
-        frame.current = 0;
-        for (let i in CANVAS.sequencepre) {
-          frame.shown[i] = 0;
-          frame.frames[i] = [i];
-        }
-      } else if (TASK.FixationUsesSample > 0) { // IF Sample, show first image/movie
-        // Update grid location of sample to current fixation grid index
-        frame.shown = [];
-        frame.frames = [];
-        frame.current = 0;
-        
-        for (let i = 0; i < CURRTRIAL.sequencegridindex.length; i++) {
-          for (let j = 0; j < CURRTRIAL.sequencegridindex[i].length; j++) {
-            if (CURRTRIAL.sequencetaskscreen == 'Sample') { // IF sample
-              // Set location to fixation
-              CURRTRIAL.sequencegridindex[i][j] = CURRTRIAL.fixationgridindex;
-
-              if (CURRTRIAL.sequenceclip[i] == 0 && j == 0) { // IF first clip, add frame
-                frame.shown.push(0);
-                frame.frames.push([i]);
-              }
-            }
-          }
-        }
-      }
-
-      // Start timer for this fixation render trial
-      CURRTRIAL.starttime = Date.now() - ENV.CurrentDate.valueOf();
-      logEVENTS('StartTime', CURRTRIAL.starttime, 'trialseries');
-
-      //========= AWAIT SHOW FIXATION =========//
-      // TODO: move to appropriate location
-      if (TASK.Species == 'marmoset' || TASK.Species == 'model') {
-        playSound(0);
-      }
-
-      if (TASK.FixationUsesSample <= 0) { // IF !FixationUsesSample, show fixation dot
-        // displayTrial(time, grid, frame, screen, obj, idx)
-        await displayTrial(
-          CANVAS.tsequencepre,
-          [CURRTRIAL.fixationgridindex],
-          [0],
-          CANVAS.sequencepre,
-          [0],
-          [0],
-        );
-      } else if (TASK.FixationUsesSample > 0) { // IF FixationUsesSample, show image/movie
-        displayTrial(
-          CURRTRIAL.tsequence,
-          CURRTRIAL.sequencegridindex,
-          CURRTRIAL.sequenceframe,
-          CURRTRIAL.sequencetaskscreen,
-          CURRTRIAL.sequencelabel,
-          CURRTRIAL.sequenceindex,
-        );
-        await moviestart_promise();
-      }
-
-      audiocontext.suspend();
-
-      //========= AWAIT HOLD FIXATION TOUCH =========//
-      if (ENV.FixationWindowRadius > 0) { // IF FixationWindow, then override object size
-        // TODO: contain the scope of funcreturn to each file.
-        funcreturn = (
-          getFixationWindowBoundingBox(
-            CURRTRIAL.fixationgridindex,
-            ENV.FixationWindowRadius
-          )
-        );
-        boundingBoxesFixation.x[0] = funcreturn[0];
-        boundingBoxesFixation.y[0] = funcreturn[1];
-      } else if (
-        TASK.FixationUsesSample > 0
-        && ENV.FixationWindowRadius <= 0
-      ) { // alt. fixation window
-        boundingBoxesFixation = boundingBoxesChoice3D;
-      }
-
-      let touchhold_return;
-      if (FLAGS.stressTest == 1) { // IF automated stress test
-        if (TASK.Species == 'model') {
-          let ctx = mkm.cvs.getContext('2d');
-          ctx.clearRect(0, 0, mkm.cvs.width, mkm.cvs.height);
-
-          touchhold_return = { type: 'theld' };
-          let sxOffset = (
-            IMAGES.Sample[CURRTRIAL.correctitem].IMAGES.sizeInches
-            * ENV.PhysicalPPI
-            / ENV.ScreenRatio
-          );
-
-          let sx = (
-            boundingBoxesFixation.x[0][1]
-            + boundingBoxesFixation.x[0][0]
-            - sxOffset
-          );
-          sx = Math.round(sx);
-
-          let syOffset = (
-            IMAGES.Sample[CURRTRIAL.correctitem].IMAGES.sizeInches
-            * ENV.PhysicalPPI 
-            / ENV.ScreenRatio
-            - ENV.FixationWindowRadius
-          );
-          let sy = (
-            (boundingBoxesFixation.y[0][1] + boundingBoxesFixation.y[0][0]) 
-            / ENV.ScreenRatio 
-            - syOffset
-          );
-          sy = Math.round(sy);
-
-          let sHeight = Math.round(
-            IMAGES.Sample[CURRTRIAL.correctitem].IMAGES.sizeInches
-            * ENV.PhysicalPPI
-          );
-          let sWidth = sHeight;
-
-          ctx.drawImage(VISIBLECANVAS, sx, sy, sWidth, sHeight, 0, 0, 224, 224);
-          let tensor = mkm.normalizePixelValues(mkm.cvs);
-          let feature = mkm.featureExtractor.execute(tensor);
-          feature = feature.reshape([2048]);
-          if (CURRTRIAL.num <= TASK.ModelConfig.trainIdx) {
-            mkm.dataObj.xTrain.push(feature);
-            if (CURRTRIAL.correctitem == 0) {
-              mkm.dataObj.yTrain.push([1, 0]);
-            } else if (CURRTRIAL.correctitem == 1) {
-              mkm.dataObj.yTrain.push([0, 1]);
-            }
-          } else {
-            mkm.dataObj.xTest = feature;
-            mkm.dataObj.yTest = CURRTRIAL.correctitem;
-          }
-
-          if (CURRTRIAL.num == TASK.ModelConfig.trainIdx) {
-            let xTrain = tf.data.array(mkm.dataObj.xTrain);
-            let yTrain = tf.data.array(mkm.dataObj.yTrain);
-            let trainDataset = tf.data.zip({ xs: xTrain, ys: yTrain })
-              .batch(4)
-              .shuffle(4);
-
-            const beginMs = performance.now();
-            await mkm.model.fitDataset(trainDataset, {
-              epochs: TASK.ModelConfig.epochs,
-              callbacks: {
-                onEpochEnd: async(epoch, logs) => {
-                  const secPerEpoch = (
-                    (performance.now() - beginMs) / (1000 * (epoch + 1))
-                  );
-                  console.log('Training model ... Approx. ' + `${secPerEpoch.toFixed(4)} sec/epoch`);
-                  console.log('logs:', logs);
-                }
-              }
-            });
-          }
-
-          let x = (
-            boundingBoxesFixation.x[0][0]
-            + Math.round(
-              Math.random() * (boundingBoxesFixation.x[0][1] - boundingBoxesFixation.x[0][0])
-            )
-          );
-
-          let y = (
-            boundingBoxesFixation.y[0][0]
-            + Math.round(
-              Math.random() * (boundingBoxesFixation.y[0][1] - boundingBoxesFixation.y[0][0])
-            )
-          );
-
-          touchhold_return.cxyt = [
-            0,
-            x,
-            y,
-            Date.now() - ENV.CurrentDate.valueOf()
-          ];
-          FLAGS.waitingforTouches--;
-        } else {
-          touchhold_return = { type: 'theld' };
-          let x = (
-            boundingBoxesFixation.x[0][0]
-            + Math.round(
-              Math.random() * (boundingBoxesFixation.x[0][1] - boundingBoxesFixation.x[0][0])
-            )
-          );
-
-          let y = (
-            boundingBoxesFixation.y[0][0]
-            + Math.round(
-              Math.random() * (boundingBoxesFixation.y[0][1] - boundingBoxesFixation.y[0][0])
-            )
-          );
-
-          touchhold_return.cxyt = [
-            0,
-            x,
-            y,
-            Date.now() - ENV.CurrentDate.valueOf()
-          ];
-
-          FLAGS.waitingforTouches--;
-        }
-      } else { // ELSE await fixation hold
-        FLAGS.acquiredTouch = 0;
-        let p1 = hold_promise(
-          TASK.FixationDuration,
-          boundingBoxesFixation,
-          FLAGS.punishOutsideTouch,
-        );
-        let p2 = choiceTimeOut(TASK.FixationTimeOut);
-        touchhold_return = await Promise.race([p1, p2]);
-      }
-
-      if (FLAGS.movieplaying == 1) { // IF movie is playing
-        // So that sample movie does not continue playing after fixation acquired
-        frame.current = frame.shown.length - 1;
-        frame.shown[frame.current] = 1;
-        await moviefinish_promise()
-      }
-
-      CURRTRIAL.fixationtouchevent = touchhold_return.type;
-      CURRTRIAL.fixationxyt = [
-        touchhold_return.cxyt[1],
-        touchhold_return.cxyt[2],
-        touchhold_return.cxyt[3],
-      ];
-		  CURRTRIAL.allfixationxyt[TASK.NFixations - FLAGS.waitingforTouches - 1] = (
-        CURRTRIAL.fixationxyt
-      );
-
-      logEVENTS("FixationTouchEvent", CURRTRIAL.fixationtouchevent, "trialseries");
-      logEVENTS("FixationXYT", CURRTRIAL.fixationxyt, "trialseries");
-      
-      // IF held fixation & fixation task, count as correct
-      if (CURRTRIAL.fixationtouchevent == 'theld') {
-        if (TASK.RewardStage == 0 && FLAGS.waitingforTouches == 0) {
-          CURRTRIAL.response = 1;
-          CURRTRIAL.correctitem = 1;
-          logEVENTS('Response', CURRTRIAL.response, 'trialseries');
-        }
-        
-        // ELSE IF broke fixation & fixation task, count as incorrect
-      } else if (TASK.RewardStage == 0 && CURRTRIAL.fixationtouchevent == 'tbroken') {
-        CURRTRIAL.response = 0;
-        CURRTRIAL.correctitem = 1;
-			  FLAGS.waitingforTouches = 0; //exit loop
-        logEVENTS("Response", CURRTRIAL.response, "trialseries");
-      }
-      /**
-         * TODO: Can probably get rid of this last else-if 
-         * ELSE IF timed out OR dms task, ok if touched outside,
-         * just wait for touch inside fixation area
-         */
-      // else if (
-      //   (CURRTRIAL.fixationtouchevent == 'tbroken'
-      //   && TASK.RewardStage == 1)
-      //   || CURRTRIAL.fixationtouchevent == 'TimeOut'
-      // ) {}
-
-      //========= AWAIT CLEAR FIXATION =========//
-      for (let q in CANVAS.sequenceblank) {
-        frame.shown[q] = 0;
-        frame.frames[q] = [q];
-      }
-      frame.current = 0;
-      if (FLAGS.waitingforTouches > 0) { // blank out screen
-        await displayTrial(
-          CANVAS.tsequenceblank,
-          [-1, -1],
-          [0, 1],
-          CANVAS.sequenceblank,
-          [0, 0],
-          [0, 0],
-        );
-      }
-    } // WHILE waiting for NFixations
-    //============ (end) WHILE RUN FIXATION SCREEN ============//
-
-
-
-    //SAMPLE TEST    SAMPLE TEST    SAMPLE TEST    SAMPLE TEST    SAMPLE TEST    //
-    //SAMPLE TEST    SAMPLE TEST    SAMPLE TEST    SAMPLE TEST    SAMPLE TEST    //
-    //SAMPLE TEST    SAMPLE TEST    SAMPLE TEST    SAMPLE TEST    SAMPLE TEST    //
-  	//============== AWAIT SHOW SAMPLE THEN TEST ==============//
-    if (TASK.RewardStage === 1) {
-
-      // Set where to display
-      if (TASK.SampleGridIndex > 0) { // IF fixed sample location
-        CURRTRIAL.samplegridindex = TASK.SampleGridIndex;
-      } else if (TASK.SampleGridIndex < 0) { // ELSE IF random sample location
-        if (TASK.FixationGridIndex < 0) { // IF moving fixation, use its grid location for sample
-          CURRTRIAL.samplegridindex = CURRTRIAL.fixationgridindex;
-        } else { // ELSE use random grid location for sample
-          CURRTRIAL.samplegridindex = (
-            Math.floor(ENV.XGridCenter.length * Math.random())
-          );
-        }
-      }
-
-      // Update grid location of each Sample frame
-      for (let i = 0; i < CURRTRIAL.sequencegridindex.length; i++) {
-        for (let j = 0; j < CURRTRIAL.sequencegridindex[i].length; j++) {
-          if (CURRTRIAL.sequencetaskscreen[i] == 'Sample') {
-            CURRTRIAL.sequencegridindex[i][j] = CURRTRIAL.samplegridindex;
-          }
-        }
-      }
-
-      logEVENTS("SampleGridIndex", CURRTRIAL.samplegridindex, "trialseries");
-      frame.shown=[];
-      frame.frames=[];
-      frame.current=0;
-      for (let q in CURRTRIAL.sequencetaskscreen) {
-        frame.shown[q] = 0;
-        frame.frames[q] = [q];
-      } // FOR q frames
-
-      // KeepSampleON
-      if (TASK.KeepSampleON == 1) {
-        let idxArr = [];
-        let idx = CURRTRIAL.sequencetaskscreen.indexOf('Sample');
-        while (idx != -1) {
-          idxArr.push(idx);
-          idx = CURRTRIAL.sequencetaskscreen.indexOf('Sample', idx + 1);
-        }
-
-        // FOR i remaining frames after Sample
-        for (let i = idxArr[idxArr.length - 1] + 1; i < frame.frames.length; i++) {
-          // Append last Sample scene rendered
-          frame.frames[i].push(idxArr[idxArr.length - 1]);
-        }
-      }
-
-      // KeepTestON
-      if (TASK.KeepTestON == 1 && TASK.SameDifferent > 0) {
-        let idxArr = [];
-        let idx = CURRTRIAL.sequencetaskscreen.indexOf('Test');
-        while (idx != -1) {
-          idxArr.push(idx);
-          idx = CURRTRIAL.sequencetaskscreen.indexOf('Test', idx + 1);
-        }
-
-        // FOR i remaining frames after TEst
-        for (let i = idxArr[idxArr.lenght - 1] + 1; i < frame.frames.length; i++) {
-          // Append last Test scene rendered
-          frame.frames[i].push(idxArr[idxArr.length - 1]);
-        }
-      }
-
-      // Display Sample & Test/Choice
-      if (TASK.NRSVP > 0 && TASK.FixationWindowSizeInches > 0) {
-        
-      }
-
-
-
-    }
-
-
-  }
-
-});
-
-
-
-
-
-
-
-
-
-(async function(){
-	document.querySelector ("button[id=quickload]").addEventListener(
-		'pointerup',quickLoad_listener,false)
-
-		//---- for Safari
-		document.querySelector ("button[id=quickload]").addEventListener(
-			'click',quickLoad_listener,false)
-
-	if ( ENV.WebUSBAvailable ){
-		await usb_scriptLoaded
-		document.querySelector ("button[id=connectusb]").addEventListener(
-			'pointerup',findUSBDevice,false)
-		document.querySelector("button[id=nousb]").addEventListener(
-			'pointerup',skipHardwareDevice,false)
-		document.querySelector("button[id=preemptRFID]").addEventListener(
-			'pointerup',preemptRFID_listener,false)	
-
-			//---- for Safari
-			document.querySelector ("button[id=connectusb]").addEventListener(
-				'click',findUSBDevice,false)
-			document.querySelector("button[id=nousb]").addEventListener(
-				'click',skipHardwareDevice,false)
-			document.querySelector("button[id=preemptRFID]").addEventListener(
-				'click',preemptRFID_listener,false)	
-			//---- (END) for Safari
-	}
-
-	if ( ENV.WebBluetoothAvailable ){
-		await ble_scriptLoaded
-		await blescale_scriptLoaded
-		//Button callback for asynchronous connection to bluetooth scale
-		document.querySelector("button[id=connectblescale]").addEventListener(
-			'pointerup',blescaleconnect,false)	
-
-			//---- for Safari
-			document.querySelector("button[id=connectblescale]").addEventListener(
-				'click',blescaleconnect,false)	
-			//---- (END) for Safari
-	}
-
-	document.querySelector("button[id=doneEditingParams]").addEventListener(
-		'pointerup',doneEditingParams_listener,false)
-	document.querySelector("button[id=doneTestingTask]").addEventListener(
-		'pointerup',doneTestingTask_listener,false)
-	document.querySelector("button[id=stressTest]").addEventListener(
-		'touchstart',stressTest_listener,false)
-	document.querySelector("button[id=gridPoints]").addEventListener(
-		'touchstart',gridPoints_listener,false)
-
-		//---- for Safari
-		document.querySelector("button[id=doneEditingParams]").addEventListener(
-			'click',doneEditingParams_listener,false)
-		document.querySelector("button[id=doneTestingTask]").addEventListener(
-			'click',doneTestingTask_listener,false)
-		document.querySelector("button[id=stressTest]").addEventListener(
-			'click',stressTest_listener,false)
-		document.querySelector("button[id=gridPoints]").addEventListener(
-			'click',gridPoints_listener,false)
-		//---- (END) for Safari
-
-
-	//====================== Retrieve device's screen properties ===========================//
-	ENV.UserAgent = window.navigator.userAgent
-	ENV.DeviceScreenWidth = window.screen.width
-	ENV.DeviceScreenHeight = window.screen.height
-
-	var deviceProperties = await deviceDetect()
-	ENV.DeviceType = deviceProperties.data.device.type
-	ENV.DeviceBrand = deviceProperties.data.device.brand
-	ENV.DeviceName = deviceProperties.data.device.model
-	ENV.DeviceGPU = deviceProperties.data.gpu.renderer
-	ENV.DeviceBrowserName = deviceProperties.data.client.name
-	ENV.DeviceBrowserVersion = deviceProperties.data.client.version
-	ENV.DeviceOSName = deviceProperties.data.os.name
-	ENV.DeviceOSVersion = deviceProperties.data.os.version
-	ENV.DeviceTouchscreen = deviceProperties.data.touchscreen
-
-	var screenSpecs = await queryDeviceonFirestore(ENV.DeviceName)
-	//if device not identified by deviceAPI or no matching firestore devices record found for an identified device
-	if (screenSpecs.screenSizeInches < 0 && ENV.DeviceType == "desktop"){
-		var screenSpecs = await queryDeviceonFirestore('32ul750') //default to desktop monitor
-		console.log('Desktop detected, defaulting to LG 32ul750 monitor for screen ppi')
-	}
-	else if (screenSpecs.screenSizeInches < 0 && ENV.DeviceType == "tablet"){
-		var screenSpecs = await queryDeviceonFirestore('pixel c') //default to pixel c
-		console.log('Tablet detected, defaulting to pixel c tablet for screen ppi')
-	}
-	else if (screenSpecs.screenSizeInches < 0 && ENV.DeviceType == "mobile"){
-		var screenSpecs = await queryDeviceonFirestore('pixel 4 xl') //default to pixel 4 xl
-		console.log('Mobile detected, defaulting to pixel 4 xl phone for screen ppi')
-	}
-	else if (screenSpecs.screenSizeInches < 0 && (ENV.DeviceType == "Not available" || ENV.DeviceType == '')){
-		var screenSpecs = await queryDeviceonFirestore('pixel c') //default to pixel c
-		console.log('Device type unidentified, defaulting to pixel c tablet for screen ppi')
-	}
-
-	ENV.ScreenSizeInches = screenSpecs.screenSizeInches
-	ENV.ScreenPhysicalPixels = screenSpecs.screenPhysicalPixels //display pixels (<= physical screen pixels)
-	ENV.ScreenRatio = screenSpecs.screenRatio //scaling from physical pixels to display pixels (retina display)
-	ENV.PhysicalPPI = screenSpecs.ppi //physical device pixels per inch
-
-	if (window.innerWidth < window.innerHeight){
-		ENV.ScreenSizeInches = [ENV.ScreenSizeInches[1],ENV.ScreenSizeInches[0],ENV.ScreenSizeInches[2]]
-		ENV.ScreenPhysicalPixels = [ENV.ScreenPhysicalPixels[1],ENV.ScreenPhysicalPixels[0]]
-	} //IF PORTRAIT flip horizontal and vertical
-
-	if (ENV.DevicePixelRatio != ENV.ScreenRatio){
-		console.log("User is not running screen at native pixelratio which affects image scaling, will attempt to compensate")
-	} //IF user not running screen at native scaling
-
-	ENV.ViewportPixels[0] = ENV.ScreenPhysicalPixels[0]/ENV.DevicePixelRatio
-	ENV.ViewportPixels[1] = ENV.ScreenPhysicalPixels[1]/ENV.DevicePixelRatio		
-
-	ENV.ViewportPPI = ENV.ViewportPixels[0]/ENV.ScreenSizeInches[0] //viewport pixels per inch
-	updateHeadsUpDisplay()
-	//====================== (END) Retrieve device's screen properties ===========================//
-
-	if ( ENV.WebUSBAvailable ){
-		var event = {}
-		event.type = "AutoConnect"
-		await findUSBDevice(event)
-	}
-
-	//====================== Quickload Button Set-up ===========================//
-	// GET PARAMFILE NAME
-	var subjectlistobj = document.getElementById("subjectID_select");
-	for (var i=subjectlist.length-1; i>=0; i--){
-		var opt = document.createElement('option');
-		opt.value = i;
-		opt.innerHTML = subjectlist[i];
-		subjectlistobj.appendChild(opt);
-	}
-	subjectlistobj.addEventListener("change",subjectlist_listener,false);
-	subjectlistobj.style.visibility = "visible"
-
-	if ( localStorage.getItem("Agent") != null ){
-		QuickLoad.agent = localStorage.getItem("Agent")
-		QuickLoad.connectusb = localStorage.getItem("ConnectUSB")
-		if (QuickLoad.connectusb == null){ QuickLoad.connectusb = 0}
-
-		document.querySelector("button[id=quickload]").style.display = "block"
-		document.querySelector("button[id=quickload]").style.visibility = "visible"
-
-		if ( QuickLoad.connectusb == 0){
-			document.querySelector("button[id=quickload]").innerHTML = QuickLoad.agent
-		}
-		else if ( QuickLoad.connectusb == 1){
-			document.querySelector("button[id=quickload]").innerHTML = QuickLoad.agent + " <i>USB</i>"
-		}
-	} //IF agent stored locally, show quickload button
-	else {
-		document.querySelector("button[id=quickload]").style.display = "none"
-	} //ELSE don't show button
-	//====================== (END) Quickload Set-up ===========================//
-
-	//================== AWAIT LOAD SUBJECT PARAMS ==================//
-	document.querySelector("div[id=subjectID_div]").style.display = "block"
-	document.querySelector("div[id=subjectID_div]").style.visibility = "visible"
-	await subjectIDPromise()	
-	document.querySelector("button[id=quickload]").style.display = "none"
-	document.querySelector("div[id=subjectID_div]").style.display = "none"
-
-	localStorage.setItem("Agent",ENV.Subject)	
-
-	ENV.ParamFileName = PARAM_DIRPATH + ENV.Subject + "_params.json";
-	await loadParametersfromFirebase(ENV.ParamFileName)
-	
-	let rtdbAgentRef = rtdb.ref('agents/' + ENV.Subject);
-	let rtdbAgentConnectionRef = rtdb.ref(`agents/${ENV.Subject}/numConnections`);
-	FLAGS.rtdbDataRef = rtdb.ref('data/' + ENV.Subject);
-	//================== (END) AWAIT LOAD SUBJECT PARAMS ==================//
-
-
-	//====================== Connect USB ===========================//
-	if ( ENV.WebUSBAvailable ){
-
-		if (typeof(port.connected) == 'undefined' || port.connected == false){
-			var event = {}
-			event.type = "AutoConnect"
-			await findUSBDevice(event)
-		}
-
-		if ( (typeof(port.connected) == 'undefined' || port.connected == false) &&
-		     (QuickLoad.load == 0 || (QuickLoad.load == 1 && QuickLoad.connectusb == 1))){
-			//=============== AWAIT CONNECT TO HARDWARE (via USB) ===============//
-			port.connected = false
-			document.querySelector("button[id=connectusb]").style.display = "block"
-			document.querySelector("button[id=connectusb]").style.visibility = "visible"
-			document.querySelector("button[id=nousb]").style.display = "block"
-			document.querySelector("button[id=nousb]").style.visibility = "visible"
-
-			await connectHardwareButtonPromise()
-		} //IF !QuickLoad.load
-
-		document.querySelector("button[id=connectusb]").style.display = "none"
-		document.querySelector("button[id=nousb]").style.display = "none"
-	}
-	else {
-		//skip usb device connection
-		port={
-		  statustext_connect: "",
-		  statustext_sent: "",
-		  statustext_received: "",
-		  connected: false
-		}
 	}
 	//====================== (END) Connect USB ===========================//
